Extract embedding storage
